+++ src/main/java/ru/practicum/shareit/item/dto/ItemDto.java
@@ -9,7 +9,7 @@ import javax.validation.constraints.NotNull;
 @Data
 @Builder
 public class ItemDto {
-    private Integer id;
+    private Long id;
 
     @NotEmpty
     private String name;
@@ -20,5 +20,5 @@ public class ItemDto {
     @NotNull
     private Boolean available;
 
-    private Integer ownerId;
+    private Long ownerId;
 }
